+++ RAM.h
@@ -48,7 +48,7 @@ public:
 
     void setLogicalRAM(int id, string mode, int depth, int width);
     string romToString(LogicalRamModes mode);
-    vector<int> mapBRAMS(int arch);
+    vector<long long> mapBRAMS(int arch);
     int  calcPhysicalBlocks(int logical_length, int physical_length);
     void saveRamMapping(long int additional_LUTs, int phys_ram_id, int p, int s, BRAMs ram_type, int phys_width, int phys_depth, int area);
     void mapLUTRAM();
